Fixed some merge errors
