+++ src/ordering/Order.java
@@ -6,31 +6,55 @@ import java.util.Set;
 import media.MediaItem;
 
 public class Order {
-	
-	
-	private int orderID;
-	private int empID;
 	private MediaItem item;
 	private String name;
 	private int quantity;
 	private String estDelivery;
+	public int ID;
 	
-	int orderIDtally = 01;
+	private static int count = 0;
 	
 	
 	
-	public Order(int orderID, int empID, MediaItem item, String name, int quantity, String estDelivery) {
-		
-		this.orderID = orderIDtally;
-		this.empID = empID;
+	public Order(MediaItem item, String name, int quantity, String estDelivery) {
 		this.item = item;
 		this.name = name;
 		this.quantity = quantity;
 		this.estDelivery = estDelivery;
 		
-		orderIDtally++;
-		
+		count++;
+		ID = count;
 	}
+	
+	/**
+	 * @return the item
+	 */
+	public MediaItem getItem() {
+		return item;
+	}
+	
+	/**
+	 * @return the name
+	 */
+	public String getName() {
+		return name;
+	}
+	
+	/**
+	 * @return the quantity
+	 */
+	public int getQuantity() {
+		return quantity;
+	}
+	
+	/**
+	 * @return the estDelivery
+	 */
+	public String getEstDelivery() {
+		return estDelivery;
+	}
+	
+
